Began working on implementing custom gear and statistics
Added the UniqueItemIdentifier class, which allows for custom items to be assigned an ID which includes a unique portion as well as a portion set to be the universal ID of the item type that item is.

Added the Quality class, which is a representation of a gear's quality.

Added the QualityEnum class, which defines the different qualities that are possible to be had by gear.

Began work on the GearItem abstract class, which will be the defualt representation of custom gear items, and will have special attributes (stats, quality) attached to them

Added the CustomStat class, which is an abstract class representative of what a custom statistic will have.

Added the StatEnum class, which contains enums for the different stat types possible.

Added the ReforgeStatistic class, which is representative of a modifiable statistic.

Added the BaseStatistic class, which is representative of a statistic native and inherent in a piece of gear, and can therefore not be modified