Player stats now update when custom gear is equipped. Added an inspect command
Player stats update when custom gear is equipped. The PlayerStat class was updated to account for the fact that is an accumulation of many different BaseStatistic and ReforgeStatistic classes. The PlayerStat's overall value is calculated by the following function: statValue = baseValue + baseValue * ((1 - (1 - percent1)) * (1 - percent2) * ... * (1 - pecentN))

Added the InspectCommandExecutor class which allows the /inspect command to be executed in game.

/inspect will display information about the statistics of the player who ran the command

/inspect <playername> will display information about the player with the specified name if they exist currently online in the server