Now require the player  to be the specified tier to upgrade generators
Added tier attribute to FFPlayer class.

Tier now shows in inspect output

when the player is not high enough tier, they can not upgrade the generator. The generator error message for upgrade failure displays the correct error (depends on if the error was not enough money or not a high enough tier)

Player can "tier up" by killing a TestChicken

Tier (just like stats) is not consistent between player sessions