Refactored Code and re-mapped spigot obfuscation to original mojang mapping
This was a necessary change, since in order to overwrite AI, you need to override the registerGoals method from an NMS entity class. With obfuscation, this was not possible, due to the composition of the obfuscated java files (could not compile due to 2 parent classes having 2 separate method declarations of x(), one returned void, and another returned boolean).

Refactored custom item code to original mojang mapping.

Continued working on implementing custom mobs