Custom Mob funtionality a
Finalized a CustomEntity interface to implement the functionality of all custom Entities

Finalized a CustomCraftEntity interface to implement the funtionality of all custom CraftEntities

Included a class to represent all custom Chicken entities, which removes all AI functionality of the chicken, and primes a function to replace it.

Implemented a HostileChickenEntity class which acts as a hostile chicken, which will attack the player.

Implemented the corresponding HostileCraftChicken and CraftChicken classes.

Implemented a PoisonChickenEntity, which extends HostileChickenEntity, has a different drop table, and applies poison effect to the player when it attacks them.